Skal ikke komme opp valget om å velge godkjent med mindre nn 75 på utsettesle + input validering + små rydding
