+++ packages/fakta-uttaksdokumentasjon/i18n/nb_NO.json
@@ -38,7 +38,7 @@
   "UttakDokumentasjonFaktaDetailForm.Avbryt": "Avbryt",
   "UttakDokumentasjonFaktaDetailForm.MorsStillingsprosent.Label": "Hvor mange prosent jobber mor?",
   "UttakDokumentasjonFaktaDetailForm.MorsStillingsprosent.ReadMoreTittel": "Når mor jobber mindre enn 75%",
-  "UttakDokumentasjonFaktaDetailForm.MorsStillingsprosent.ReadMoreInnhold": "Når mor arbeider deltid, blir fars utbetaling av foreldrepenger redusert tilsvarende reduksjonen i hennes arbeidstid. Far får dermed lavere utbetaling, men hans periode blir ikke forlenget som ved gradering. Hvis mor for eksempel jobber 50 prosent, får far utbetalt 50 prosent av sin fulle dagsats. Hvis mor jobber 75 prosent av en full stilling eller mer, får likevel far utbetalt foreldrepenger etter sin fulle dagsats. Mer om <a>vilkår for fars uttak (åpnes i ny fane)</a>.",
+  "UttakDokumentasjonFaktaDetailForm.MorsStillingsprosent.ReadMoreInnhold": "Når mor arbeider deltid under 75 prosent, blir fars utbetaling av foreldrepenger redusert tilsvarende mors stillingsprosent, og han får fulle trekkdager. Dersom han tar ut mindre foreldrepenger enn mors stillingsprosent vil han få utbetalt det han søker om, og antall trekkdager blir redusert i henhold til mors stillingsprosent (slik som ved gradering). Mer om <a>vilkår for fars uttak (åpnes i ny fane)</a>.",
   "UttakDokumentasjonFaktaDetailForm.Vurdering": "Vurdering",
   "UttakDokumentasjonFaktaDetailForm.Fom": "Fra og med:",
   "UttakDokumentasjonFaktaDetailForm.Tom": "Til og med:",
@@ -47,7 +47,7 @@
   "UttakDokumentasjonFaktaDetailForm.DelOppPeriode": "Del opp periode",
   "UttakDokumentasjonFaktaDetailForm.IkkeGyldigDato": "Dato er utenfor opprinnelig periode",
   "UttakDokumentasjonFaktaDetailForm.Godkjent": "Godkjent",
-  "UttakDokumentasjonFaktaDetailForm.GodkjentOver75": "Godkjent - Mor jobber mer enn 75%",
+  "UttakDokumentasjonFaktaDetailForm.GodkjentOver75": "Godkjent - Mor jobber mer eller lik 75%",
   "UttakDokumentasjonFaktaDetailForm.GodkjentUnder75": "Godkjent - Mor jobber mindre enn 75%",
   "UttakDokumentasjonFaktaDetailForm.IkkeGodkjent": "Ikke godkjent",
   "UttakDokumentasjonFaktaDetailForm.IkkeDokumentert": "Mangler dokumentasjon",
