+++ packages/fakta-uttaksdokumentasjon/types/FormValues.ts
@@ -1,5 +1,9 @@
 import { DokumentasjonVurderingBehov } from '@navikt/fp-types';
 
+type FormValues = {
+  perioder: VurderingsBehovPeriode[];
+};
+
 export enum VurderingsAlternativ {
   GODKJENT = 'GODKJENT',
   GODKJENT_OVER75 = 'GODKJENT_OVER75',
@@ -11,3 +15,5 @@ export enum VurderingsAlternativ {
 export type VurderingsBehovPeriode = Omit<DokumentasjonVurderingBehov, 'vurdering'> & {
   vurdering?: VurderingsAlternativ;
 };
+
+export default FormValues;
